Fix mouse confinement being toggled on click
This commit makes it so that SDL_CaptureMouse is used instead of
SDL_SetGrabWindow to capture mouse up events that occur outside the SDL
window.

The existing comment makes it clear that this is what the original
developers intended, but the pre-existing code actually performed "don't
let the mouse cursor leave the window until it is released".